Lab Task 12/requirements.txt
flask
pandas
numpy
sentence-transformers
faiss-cpu